chore: improve Makefile for minimal POSIX shell compatibility
When the shell used for `sh` is `dash`, which is common on Debian and
its derivatives, the output of the Makefile is incorrect. This patch
resolves the issue by using `bash` as the shell when it's available.

Additionally, I refactored the `push` target to reuse the code that
executes the push command for a specific directory.

Lastly, the default action when you run `make` without any arguments
is now to build all the extensions.

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>